requirements.txt
python-dotenv==0.19.0
numpy
Flask==2.2.0
urllib3==1.26.12
python-telegram-bot==20.3
telegram==0.0.1
openai==0.27.6
pandas
tiktoken
matplotlib
plotly==5.15.0
scipy
scikit-learn
typing-extensions
langchain==0.0.199
CairoSVG==2.7.0